requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.5
django-autoslug==1.9.8
django-crispy-forms==1.9.2
django-etc==1.2.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-markdown-deux==1.0.5
djongo==1.3.3
dnspython==2.0.0
idna==2.10
markdown2==2.3.9
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.2.4
urllib3==1.25.10
whitenoise==5.2.0
